fix: move custom mapping to a separate file
